Quick overview of Bootstrap Grid Layout and Bootstrap Components
Briefing of Bootstrap and learning how it works in VS code and viewing it on the Chrome browser Responsive in various scales (ie. desktop, mobile, tablet, etc) Utilizing codes to help build a website with getbootstrap.com. Used Bootstrap Components: Navbar, Carousel, Jumbotron and Cards. Input them with the appropriate image for this exercise.