Learned to customize the default Bootstrap 4 css styles
Did a small exercise to understand how to create container with bootstrap and styling it. Applied to the html. Added the css link and ensure the link goes AFTER the bootstrap link to be able to over ride the bootstrap if want to change certain things from Bootstrap. Created css to modify some styles such as font and adding gradient background. Also, added body font to the webpage. Used Google Font to find the font-family and inplement it in the html and css pages.