Bootstrap Exercise - Turn things on and off, re-create Bootstrap media queries
Lessons to utilize website: getbootstrap.com to find the necessary coding to adjust the website using Bootstrap 4. Turned things on and off when the website has been resized/responsive from mobile to desktop. Also, re-create the Bootstrap media queries in own CSS file. Used colors to showcase the changes in the media queries in the CSS.