Google Chrome Inspect - Removing Overriding Bootstrap styles
Learned to inspect and find certain styles and how to change them for viewing purposes in inspect. Also, when using Bootstrap to not to try to override the styles with css too much but with Bootstrap itself. Used the getbootstrap to retrieve certain bootstrap codes to override styles properly such as change h1 text to white by adding the text-white under the Jumbotron for "Hello, world!".