Add text drop shadow and boxes - when to use boostrap code and when to create your own style
What Bootstrap does and doesn't and what to input yourself to create the style you want like the text drop shadow and lead. When you cannot find the bootstrap style for it on getbootstrap.com, you will eventually have to make your own style in css. In this exercise, changed out the h1 text, added text drop shadow to h1 only, updated lead text in css and for the button, changed the text itself to Pricing.