Change default Bootstrap 4 button size & color, images responsive stretchy in BS4
Changed the style of the button to be outline white and lightly colored hover using the btn-outline-light and adding padding px, py for the button.  Added image to make it img-fluid so it is placed on the right hand side of the h1 as desired.  Modified the image so it responds to the different sizes by using col-12 col-md-5.