Insert Logo, removed part of the nav links and created buttons
Inserted the logo - the SVG logo didn't appear as it should from the exercise files (but we get it's there!!). Also, relocated the "dropdown" <li> to go before the "Contact" button by placing it before the <li> of contact. Made it same button as the "pricing" with outline and text white. Added padding between the buttons and also added a list for the dropdown menu. Mostly done in the html file as bootstrap seems to like it there than in css file. Removed all the bg-light so the logo doesn't have a box around and behind it.