Added image, text align center and rounded corners on one side
Added the found image and the message from founder to the website. Made the message on it's own "section" so it can be modified separately from the image. Border was added to the message box and rounded the one side of the outside corners of the image and messgae box. Used mostly bootstrap styling in html to make it work Bootstrap way.