Bootstrap to style cards section
Selected the cards section and added a wrap to be card group. Then changed the "card-group" to "card-deck" with my-5 (padding around each of it) so there is space in between each card and added back shadows for each one by adding "shadow" to each of the class="card". This way the shadow will shadow the cards only and not as a box if you were to shadow with  the "card-deck" class. Also, look at getbootstrap for other ways to get the style you want under cards.